Compilation flag to fully expose libclang
To facilitate libclang extension, fully expose the library symbols if
explicitely requested.

TN: U208-020
Change-Id: Ia7e2af33667562473ae5988d1a7e0b772decce74

Reenable the symbol export list for shared libclang

We don't use the shared version of the library, and the lack of an
export list causes a linker error with LLVM 19 because the library
exports too many symbols. The static version of the library, used by
GNATcoverage, isn't affected by this change.

Issue: eng/toolchain/llvm-project#17