Fix issue with TBAA tags on aggregate objects
When LLVM checks to see if a scalar and aggregate have conflicting TBAA
access tags, it checks subfields to see if the actually match the scalar
type, not if they conflict.  This patch is temporary, but is the simplest
patch to fix the issue.  It will be raised on the LLVM development list.