Fix rewriting RemoveText API
When removing a source range from the RewriterBuffer, clang needs to
compute the size of the actual range in the RewriterBuffer (as new
elements may have been inserted at any position of the original range).
Once this is done, it then maps the original source location offset to
real offsets into the RewriteBuffer, supposedly accounting for
modifications of the RewriteBuffer (insertion / deletion of text).
Nevertheless, it actually does not acount for modifications at the
beginning of the range, which is unconsistent behavior as they are taken
into account when computing the size of the actual range. This commit
fixes this slightly-off behavior.

See merge requests eng/toolchain/llvm-project!3,
eng/toolchain/llvm-project!4, and eng/toolchain/llvm-project!11.