Two fixes for DISubrangeType
While implementing more support for subranges, I unfortunately found a
couple of bugs in my previous patch for DISubrangeType.  First, when
using a DISubrangeType as an array index type, the wrong tag was
written into the DIE.  Second, I'd intended for subranges to use bit
strides, not byte strides -- but neglected to actually implement this.