Two DWARF variant part improvements
This patch adds a couple of improvements to the LLVM emission of DWARF
variant parts.  One of these is desirable for Ada, and the other is
required.

Currently, when emitting a discriminant, LLVM follows the precise
letter of the DWARF standard, which says:

    If the variant part has a discriminant, the discriminant is
    represented by a separate debugging information entry which is a
    child of the variant part entry.

However, for Ada this does not really make sense.  In Ada, the
discriminant field exists outside of any variant part, and it makes
more sense to emit it separately rather than redundantly emit the
field once for each variant part.

This extension was arrived at when this was implemented in GCC, and
was accepted for DWARF 6, see:

    https://dwarfstd.org/issues/180123.1.html

Here the patch simply lifts this restriction: if the discriminant
field was already emitted, it isn't re-emitted.  This approach allows
the Ada compiler to do what it needs without affecting the Rust
output.

Second, this patch extends the discriminant to allow multiple values.
This is needed by Ada.  Here, I chose to use a ConstantDataArray of
pairs of integers, with each pair representing a range, as Ada also
allows ranges here.  This seemed like a reasonably convenient
representation.