Omit member size from DWARF when desired
It's reasonably normal for DWARF not to explicitly state the size of a
member of a composite type -- in this case the size is taken from the
type.  LLVM does allow a member to have a NULL size, but in this case
it emits a size of "0".  This patch cleans this up a bit, changing the
DWARF emission code to simply not emit a size in this situation.